+++ crates/miden-protocol/src/block/mod.rs
@@ -16,8 +16,8 @@ pub use block_outputs::BlockOutputs;
 mod proposed_block;
 pub use proposed_block::ProposedBlock;
 
-mod validator_keys;
-pub use validator_keys::{ValidatorKeys, ValidatorKeysError};
+mod validator_config;
+pub use validator_config::ValidatorConfig;
 
 mod block_signatures;
 pub use block_signatures::{BlockSignatures, SignatureVerificationError};